make search string used to look up objects configurable (#884)
* make search string used to look up objects configurable

* fix lint require for ci, add example to initializer

* fix typo

* move object factory specs to all be in one place

* make rubocop happy

* unpin dry-monads. its not a dependency of bulkrax, but previously needed to be pinned for specs

---------
